requirements.txt
requests
pandas
python-telegram-bot
python-dateutil
pandas-ta
setuptools
websocket-client
matplotlib